Refactor & add register event callbacks
