+++ pyown/items/energy/__init__.py
@@ -1,2 +1,4 @@
-from .energy import EnergyManagement, TypeEnergy
-from .stop_go import StopGo, StopGoStatus
+from .dataclass import StopGoStatus
+from .energy import EnergyManagement
+from .stop_go import StopGo
+from .values import TypeEnergy
